+++ requirements.txt
@@ -1,2 +1,33 @@
 spacy>=3.0.0
-num2words
+num2wordsaltgraph==0.17.3
+appJar==0.94.0
+asn1crypto==1.5.1
+cobble==0.1.3
+cymem==2.0.10
+dnspython==2.7.0
+docopt==0.6.2
+greenlet==3.0.3
+mammoth==1.5.0
+pefile==2023.2.7
+pg8000==1.31.2
+Pillow==9.4.0
+psycopg2-binary==2.9.9
+Pygments==2.19.1
+pyinstaller==5.8.0
+pyinstaller-hooks-contrib==2023.0
+pymongo==4.10.1
+python-dateutil==2.9.0.post0
+pywin32-ctypes==0.2.0
+scramp==1.4.5
+shellingham==1.5.4
+six==1.16.0
+soupsieve==2.3.2.post1
+spacy-legacy==3.0.12
+spacy-loggers==1.0.5
+SQLAlchemy==2.0.31
+sqlparse==0.4.4
+typing_extensions==4.12.2
+tzdata==2024.1
+urllib3==2.3.0
+wrapt==1.17.1
+XlsxWriter==3.2.0
