+++ src/mathspell/__init__.py
@@ -1,1 +1,2 @@
 from .main import analyze_text
+from .helpers import *
